Cache patient data for 5 mins
This is mainly important for subsequent API calls during app load. The patient instance is created to fetch patient data for display, then for trial data and again for the patient photo. To avoid making a round-trip to the SMART server three times within a few seconds, patient data is cached in the MongoDB. Cache data is discarded if it's older than 5 minutes.